+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+* PhysioNet 2012 challenge time stamp is now in hours (previously minutes) for consistency with the PhysioNet 2019 challenge
+
 ## [0.5.1] - 2022-08-03
 
 ### Changed